Version 25.9
- Added `ContentLightLevelInfo` box parsing;
- Added access to more metadata (including ColourType, GroupsListBox, etc.);
- Added `toString` summary method to image frame;
- Improved parallel processing for enhanced stability;
- Changed the way `Frames` property is calculated (altr group logic and no thumbnails);
- Changed frame properties loading to support iOS18+ files;
- Added box types used in iOS18+ files;
- Performed memory optimizations;
- API markdown fixes;
- Minor spelling fixes.

+++ src/main/java/openize/heic/decoder/coding_tree_unit.java
@@ -25,7 +25,7 @@ class coding_tree_unit
         int xCtb = (int)((header.getCtbAddrInRs() & 0xFFFFFFFFL) % (sps.getPicWidthInCtbsY() & 0xFFFFFFFFL)) << (sps.getCtbLog2SizeY() & 0xFF);
         int yCtb = (int)((header.getCtbAddrInRs() & 0xFFFFFFFFL) / (sps.getPicWidthInCtbsY() & 0xFFFFFFFFL)) << (sps.getCtbLog2SizeY() & 0xFF);
 
-        picture.SliceAddrRs[xCtb >> (sps.getCtbLog2SizeY() & 0xFF)][yCtb >> (sps.getCtbLog2SizeY() & 0xFF)] = header.getSliceAddrRs();
+        picture.SliceAddrRs[xCtb >> (sps.getCtbLog2SizeY() & 0xFF)][yCtb >> (sps.getCtbLog2SizeY() & 0xFF)] = (int)header.getSliceAddrRs();
         picture.SliceHeaderIndex[xCtb][yCtb] = header.slice_index;
 
         if (header.slice_sao_luma_flag || header.slice_sao_chroma_flag)
